+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 ## 1.5.0
+* Use PlatformDispatcher.instance over window.
+* Use super parameters lint.
+* Add Multiselect Dropdown with Checkboxes Example.
+* Update README.
+
+## 1.5.0
 ### Breaking Changes:
 * Flutter 3.0.0 upgrade.
 * Update README.